Model Deployment/Yolo Deployment/requirements.txt
absl-py==1.0.0
altair==4.2.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
bleach==4.1.0
blinker==1.4
cachetools==5.0.0
certifi==2021.5.30
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
cycler==0.10.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.3
fonttools==4.31.2
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.6.2
google-auth-oauthlib==0.4.6
grpcio==1.44.0
idna==2.10
importlib-metadata==4.11.3
importlib-resources==5.6.0
ipykernel==6.12.1
ipython==8.2.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
jedi==0.18.1
Jinja2==3.1.1
jsonschema==4.4.0
jupyter-client==7.2.1
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.1.0
kiwisolver==1.3.1
Markdown==3.3.6
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.5.13
nbconvert==6.4.5
nbformat==5.3.0
nest-asyncio==1.5.5
notebook==6.4.10
numpy==1.22.3
oauthlib==3.2.0
opencv-python-headless==4.5.5.64
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.1.0
prometheus-client==0.14.0
prompt-toolkit==3.0.29
protobuf==3.20.0
psutil==5.9.0
pure-eval==0.2.2
pyarrow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydeck==0.7.1
Pygments==2.11.2
Pympler==1.0.1
pyparsing==2.4.7
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==22.3.0
requests==2.27.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
roboflow==0.2.4
rsa==4.8
scipy==1.8.0
seaborn==0.11.2
semver==2.13.0
Send2Trash==1.8.0
six==1.16.0
smmap==5.0.0
soupsieve==2.3.2
stack-data==0.2.0
streamlit==1.8.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
terminado==0.13.3
testpath==0.6.0
thop==0.0.31.post2005241907
toml==0.10.2
toolz==0.11.2
torch==1.11.0
torchvision==0.12.0
tornado==6.1
tqdm==4.64.0
traitlets==5.1.1
typing_extensions==4.1.1
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.6
validators==0.18.2
watchdog==2.1.7
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.1.1
wget==3.2
widgetsnbextension==3.6.0
zipp==3.8.0